Move reviewTaskMachineProvider into machine folder - We chose not to use Jotai as the provider for this machine because we need to recreate the machine service every time the drawer is opened and at the moment the best option seems to be to use context rather than try to get around Jotai's default behavior of keeping the first machine created
