+++ src/components/SentenceReport/SentenceReportView.tsx
@@ -2,7 +2,7 @@ import { useSelector } from "@xstate/react";
 import { Col, Row } from "antd";
 import React, { useContext } from "react";
 
-import { ReviewTaskMachineContext } from "../../Context/ReviewTaskMachineProvider";
+import { ReviewTaskMachineContext } from "../../machines/reviewTask/ReviewTaskMachineProvider";
 import { Roles } from "../../types/enums";
 import {
     crossCheckingSelector,
